+++ src/JobInterface.php
@@ -16,6 +16,7 @@ interface JobInterface
 {
     /**
      * @param Queue $queue which pushed and is handling the job
+     * @return void|mixed result of the job execution
      */
     public function execute($queue);
 }