+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Yii2 Queue Extension Change Log
 - Bug #258: Worker in isolated mode fails if PHP_BINARY contains spaces (luke-)
 - Bug #269: Handling of broken messages that are not unserialized correctly (zhuravljov)
 - Bug #267: Fixed symfony/process incompatibility (rob006)
+- Enh: Job execution result that forwarded to the event handler (zhuravljov)
 
 2.1.0 May 24, 2018
 ------------------